backend/requirements.txt
fastapi
uvicorn
pymysql
pydantic
sqlalchemy
python-dotenv
passlib[bcrypt]